refactor: Extract out action handler
