+++ app/configurator/configurator-state/reducer.tsx
@@ -54,6 +54,7 @@ import { getChartConfig, makeMultiFilter } from "@/config-utils";
 import { mapValueIrisToColor } from "@/configurator/components/ui-helpers";
 import {
   addDatasetInConfig,
+  removeDatasetInConfig,
   getPreviousState,
   getStateWithCurrentDataSource,
   hasChartConfigs,
@@ -1083,61 +1084,7 @@ const reducer_: Reducer<ConfiguratorState, ConfiguratorStateAction> = (
 
     case "DATASET_REMOVE":
       if (isConfiguring(draft)) {
-        const { locale } = action.value;
-        const chartConfig = getChartConfig(draft);
-        const removedCubeIri = action.value.iri;
-        const newCubes = chartConfig.cubes.filter(
-          (c) => c.iri !== removedCubeIri
-        );
-        const dataCubesComponents = getCachedComponents({
-          locale,
-          dataSource: draft.dataSource,
-          cubeFilters: newCubes.map((cube) => ({
-            iri: cube.iri,
-            joinBy: newCubes.length > 1 ? cube.joinBy : undefined,
-          })),
-        });
-
-        if (!dataCubesComponents) {
-          throw Error(
-            "Error while removing dataset: Could not find cached dataCubesComponents"
-          );
-        }
-
-        const { dimensions, measures } = dataCubesComponents;
-        const iris = chartConfig.cubes
-          .filter((c) => c.iri !== removedCubeIri)
-          .map(({ iri }) => ({ iri }));
-        const { enabledChartTypes } = getEnabledChartTypes({
-          dimensions,
-          measures,
-          cubeCount: iris.length,
-        });
-        const initialConfig = getInitialConfig({
-          chartType: enabledChartTypes.includes(chartConfig.chartType)
-            ? chartConfig.chartType
-            : enabledChartTypes[0],
-          iris,
-          dimensions,
-          measures,
-          meta: current(chartConfig.meta),
-        });
-        const newChartConfig = deriveFiltersFromFields(initialConfig, {
-          dimensions,
-        });
-        const initConfig = getInitialConfiguringConfigBasedOnCube({
-          dataSource: draft.dataSource,
-          chartConfig: newChartConfig,
-        });
-        const newConfig = {
-          ...initConfig.chartConfigs[0],
-          key: chartConfig.key,
-        } as ChartConfig;
-        const index = draft.chartConfigs.findIndex(
-          (d) => d.key === chartConfig.key
-        );
-        const withFilters = deriveFiltersFromFields(newConfig, { dimensions });
-        draft.chartConfigs[index] = withFilters;
+        removeDatasetInConfig(draft, action.value);
 
         return draft;
       }
